separa componentes
